Promote SessionChrome to production session dock.
Wire the former Morph dock into the real session route, rename exports/tokens to neutral SessionChrome, collapse the composer-chrome gallery to a scenario harness, and keep queue draft-replace plus token tooltips.
