+++ apps/web/src/routes/composer-chrome-gallery.tsx
@@ -1,2 +1,2 @@
-// Back-compat re-export — gallery lives under `@/routes/composer-chrome`.
+// Back-compat re-export — harness lives under `@/routes/composer-chrome`.
 export { ComposerChromeGalleryRoute } from "./composer-chrome";
