+++ apps/web/src/routes/composer-chrome/scenario-stack.tsx
@@ -1,12 +1,15 @@
-// Faithful production chrome stack for the DEV gallery.
-// Same order as SessionChat in session.tsx — do not redesign.
-import { ChatComposer, QueueSurface, type ComposerState } from "@opengeni/react";
-import type { ClientVoiceInputConfig } from "@opengeni/sdk";
+// Production SessionChrome + ChatComposer stack for the DEV harness.
+import {
+  ChatComposer,
+  SessionChrome,
+  type ComposerState,
+  type UseTurnQueueResult,
+} from "@opengeni/react";
+import type { ClientVoiceInputConfig, SessionPendingInputPreview, SessionTurn } from "@opengeni/sdk";
 import { useMemo, useState } from "react";
 
 import { ModelPicker, SessionToolPicker } from "@/components/pickers";
-import { ComposerAgentsPill } from "@/components/session/composer-agents-pill";
-import { GoalSurface } from "@/components/session/goal-surface";
+import { SubagentTree } from "@/components/session/subagents";
 import {
   emptyAttachments,
   galleryFirstPartyTools,
@@ -31,6 +34,69 @@ const fixtureClient = {
   },
 };
 
+/** Local mutable queue/inbox so harness hover actions visibly update. */
+function useHarnessLiveQueue(seed: UseTurnQueueResult): {
+  queue: UseTurnQueueResult;
+  dismissIncoming: (inputId: string) => void;
+} {
+  const [turns, setTurns] = useState<SessionTurn[]>(seed.queue);
+  const [inputs, setInputs] = useState<SessionPendingInputPreview[]>(seed.pendingInputs);
+
+  const queue = useMemo<UseTurnQueueResult>(
+    () => ({
+      ...seed,
+      queue: turns,
+      pendingInputs: inputs,
+      moveTurn: async (turnId, beforeTurnId) => {
+        setTurns((prev) => {
+          const from = prev.findIndex((turn) => turn.id === turnId);
+          if (from < 0) return prev;
+          const next = [...prev];
+          const [moved] = next.splice(from, 1);
+          if (!moved) return prev;
+          if (beforeTurnId === null) {
+            next.push(moved);
+          } else {
+            const to = next.findIndex((turn) => turn.id === beforeTurnId);
+            if (to < 0) next.push(moved);
+            else next.splice(to, 0, moved);
+          }
+          return next;
+        });
+        return true;
+      },
+      editTurn: async (turnId) => {
+        setTurns((prev) => prev.filter((turn) => turn.id !== turnId));
+        return null;
+      },
+      steerTurn: async (turnId) => {
+        setTurns((prev) => {
+          const from = prev.findIndex((turn) => turn.id === turnId);
+          if (from <= 0) return prev;
+          const next = [...prev];
+          const [moved] = next.splice(from, 1);
+          if (!moved) return prev;
+          next.unshift(moved);
+          return next;
+        });
+        return true;
+      },
+      removeTurn: async (turnId) => {
+        setTurns((prev) => prev.filter((turn) => turn.id !== turnId));
+        return true;
+      },
+    }),
+    [inputs, seed, turns],
+  );
+
+  return {
+    queue,
+    dismissIncoming: (inputId) => {
+      setInputs((prev) => prev.filter((input) => input.id !== inputId));
+    },
+  };
+}
+
 export function ScenarioStack({
   scenario,
   composer,
@@ -42,21 +108,54 @@ export function ScenarioStack({
   const [effort, setEffort] = useState<IntelligenceEffort>("medium");
   const [toolSelection, setToolSelection] = useState(galleryToolSelection);
   const attachments = useMemo(() => emptyAttachments(), []);
+  const { queue, dismissIncoming } = useHarnessLiveQueue(scenario.queue);
+  const agents = scenario.agentNodes;
+
+  const runningAgents = agents.filter(
+    (node) => node.session.status === "running" && node.session.effectiveControl.state === "active",
+  ).length;
+  const pausedAgents = agents.filter((node) => node.session.effectiveControl.state === "paused")
+    .length;
 
   return (
     <div
-      className="flex flex-col justify-end rounded-xl border border-border bg-bg/40 pt-10"
+      className="flex flex-col justify-end rounded-xl border border-border bg-bg/40 pt-8"
       data-scenario={scenario.id}
+      data-session-chrome-stack=""
     >
-      <QueueSurface queue={scenario.queue} composer={composer} />
-      <GoalSurface session={scenario.session} goal={scenario.goal} />
-      <ComposerAgentsPill workspaceId={GALLERY_WORKSPACE_ID} nodes={scenario.agentNodes} />
+      <div className="mx-auto mb-2 w-full max-w-3xl shrink-0 px-4 sm:px-6">
+        <SessionChrome
+          queue={queue}
+          composer={composer}
+          goal={scenario.goal}
+          onDismissIncoming={dismissIncoming}
+          agentsSignal={
+            agents.length > 0
+              ? {
+                  count: agents.length,
+                  detail:
+                    runningAgents > 0
+                      ? `${runningAgents} running`
+                      : pausedAgents > 0
+                        ? `${pausedAgents} paused`
+                        : "Idle",
+                  tone: runningAgents > 0 ? "running" : pausedAgents > 0 ? "waiting" : "neutral",
+                }
+              : undefined
+          }
+          agentsPanel={
+            agents.length > 0 ? (
+              <SubagentTree workspaceId={GALLERY_WORKSPACE_ID} nodes={agents} />
+            ) : null
+          }
+        />
+      </div>
       <div className="shrink-0 px-4 pb-4 pt-1 sm:px-6">
         <div className="mx-auto w-full max-w-3xl">
           <ChatComposer
             composer={composer}
             effectiveControl={scenario.session.effectiveControl}
-            queuedAheadCount={scenario.queue.queue.length}
+            queuedAheadCount={queue.queue.length}
             placeholder="Send a follow-up…"
             attachments={attachments}
             transcription={{
