+++ docs/architecture.md
@@ -417,7 +417,7 @@ The monorepo is a **Bun workspace** (`workspaces: [apps/*, examples/*, packages/
 | `deployment`    | `@opengeni/deployment`    | Deployment contract source of truth: 12 profiles, overlays, required-env/preflight/stack-plan/runtime-artifact logic.                                                                                                                                                                                                                                                                               | `contractForProfile`, `preflightChecksFor`, `stackPlanFor`, `generateRuntimeArtifacts`, `SANDBOX_REQUIRED_ENV` (parity-pinned to config).                                  | `packages/deployment/src/index.ts`                                                                        |
 | `agent-proto`   | `@opengeni/agent-proto`   | Generated TS wire types for the selfhosted control protocol, codegen'd from `agent/proto/opengeni_agent.proto`. Control-plane side of the BYO-compute seam.                                                                                                                                                                                                                                         | Re-exports the generated `opengeni.agent.v1` types.                                                                                                                        | `packages/agent-proto/src/index.ts`                                                                       |
 | `sdk`           | `@opengeni/sdk`          | **Published.** Zero-runtime-dep TS client: typed API methods + exactly-once SSE streaming + proxy helpers + desktop/terminal transport contracts plus native `transcribeAudio` (legacy adapter types deprecated). | `OpenGeniClient`, `streamSessionEvents`, `transcribeAudio`, proxy/desktop/terminal helpers. | `packages/sdk/src/client.ts`, `packages/sdk/src/stream.ts`, `packages/sdk/src/transcription.ts`            |
-| `react`         | `@opengeni/react`        | **Published.** React hooks + styled components over the SDK: live streaming, chat, timeline, sandbox surfaces, and native MediaRecorder composer voice input. Product hosts can use the session-only **`@opengeni/react/session`** subpath for session hooks and pure projection without the styled workbench graph; its structural client type excludes workspace administration and workbench APIs. Advanced composer composition lives under **`@opengeni/react/composer`**; Connected-machine UI lives under **`@opengeni/react/machines`**. Opt-in compact session chrome is **`MorphSessionChrome`** (token-themed `--og-morph-*`; production may keep `QueueSurface` until switched). | `OpenGeniProvider`, the `use*` hooks, `ChatComposer`/`MorphSessionChrome`/`MessageTimeline`/`WorkspaceDock`/etc.; `./session` for host-rendered session embedding; `./session-ui` for styled session surfaces; `./composer` for composable composer UI; `./machines` for machine UI. | `packages/react/src/index.ts`, `packages/react/src/session.ts`, `packages/react/src/session-ui.ts`, `packages/react/src/composer.ts`, `packages/react/src/components/morph-session-chrome.tsx`, `packages/react/src/hooks/use-voice-input.ts`, `packages/react/src/machines.ts`, `packages/react/src/client.ts` |
+| `react`         | `@opengeni/react`        | **Published.** React hooks + styled components over the SDK: live streaming, chat, timeline, sandbox surfaces, and native MediaRecorder composer voice input. Product hosts can use the session-only **`@opengeni/react/session`** subpath for session hooks and pure projection without the styled workbench graph; its structural client type excludes workspace administration and workbench APIs. Advanced composer composition lives under **`@opengeni/react/composer`**; Connected-machine UI lives under **`@opengeni/react/machines`**. Built-in compact session chrome is **`SessionChrome`** (token-themed `--og-session-chrome-*`; merges incoming/queue/goal/agents above the composer). `QueueSurface` remains available for embeds that want the standalone queue only. | `OpenGeniProvider`, the `use*` hooks, `ChatComposer`/`SessionChrome`/`MessageTimeline`/`WorkspaceDock`/etc.; `./session` for host-rendered session embedding; `./session-ui` for styled session surfaces; `./composer` for composable composer UI; `./machines` for machine UI. | `packages/react/src/index.ts`, `packages/react/src/session.ts`, `packages/react/src/session-ui.ts`, `packages/react/src/composer.ts`, `packages/react/src/components/session-chrome.tsx`, `packages/react/src/hooks/use-voice-input.ts`, `packages/react/src/machines.ts`, `packages/react/src/client.ts` |
 | `ogtool`        | `@opengeni/ogtool`        | **Published.** Dependency-free Toolspace CLI for stock sandboxes, custom rigs, and connected machines. Its package bin is the canonical source installed into OpenGeni images and may be bootstrapped only from an exact deployment-pinned stable package spec. | `ogtool list`, `ogtool call`, `ogtool doctor`; package constants for the CLI environment contract. | `packages/ogtool/bin/ogtool.cjs`, `packages/ogtool/src/index.ts`                                          |
 | `testing`       | `@opengeni/testing`       | Shared test harness/fixtures (`startTestServices`, `buildSandboxImage`, `ScriptedModel`, e2e worker).                                                                                                                                                                                                                                                                                               | Consumed only by tests.                                                                                                                                                    | `packages/testing`                                                                                        |
