+++ packages/react/CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ## Unreleased
 
-- Add `MorphSessionChrome` — opt-in compact session-signal dock (incoming / queue / goal / agents) with token overrides (`--og-morph-*`) and queue hover actions wired to `UseTurnQueueResult`.
+- Add `SessionChrome` — production/embed session-signal dock (incoming / queue / goal / agents) with token overrides (`--og-session-chrome-*`), token tooltips, crossfade panel switches, and queue hover actions wired to `UseTurnQueueResult`. Replaces stacked `QueueSurface` + host goal/agents chrome in the web session route.
+- Export shared `Tooltip*` primitives used by SessionChrome (portal-token aware).
 - Light ChatComposer spacing/token freshen (quieter focus border, tighter footer/input padding).
 - Replace host-adapter composer transcription with native MediaRecorder capture and immediate server transcription.
 - Compact composer voice chrome: live/fallback waveform, separate cancel vs stop-and-transcribe actions, and a quieter transcribing state.
